Move map env expansion to util/env
